+++ CHANGELOG.md
@@ -31,6 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - fix(asea): fixed mapping upload logic to use lza enabled regions and accounts
 - fix(networking): fixed hosted zones for sagemaker vpc endpoints
 - fix(service quota): fix region logic that provisioned service limit for home region even if not specified
+- fix(servicecatalog): fixed permission set role matching logic
 
 ### Changed